requirements.txt
Django
django-extensions
ipython
django-admin-rangefilter
django-import-export
celery
django-filter
djangorestframework
drf-excel

# dev
# TODO check view name is not displayed
django-debug-toolbar

# flake8
flake8==4.0.1
flake8-debugger==4.1.2
flake8-print==5.0.0
flake8-builtins==1.5.3
